Refactor options dialog
